refactor(habitat): extract hook procedure context policy
